Implemented product module and removed product state from picklist module (#5cwacm)
