+++ src/store/modules/picklist/PicklistState.ts
@@ -1,7 +1,5 @@
 export default interface PicklistState {
-  cached: any;
   current: object;
   list: {};
-  products: {};
   selectedProducts: any[];
 }
