+++ src/store/modules/picklist/mutation-types.ts
@@ -2,4 +2,3 @@ export const SN_PICKLIST = 'picklist'
 export const PICKLIST_CURRENT = SN_PICKLIST + '/CURRENT'
 export const PICKLIST_ITEMS = SN_PICKLIST + '/ITEMS'
 export const PICKLIST_SELECTED_PRODUCTS = SN_PICKLIST + '/SELECTED_PRODUCTS'
-export const PRODUCT_ADD_TO_CACHED_MULTIPLE = SN_PICKLIST + '/ADD_TO_CACHED_MULTIPLE'
