+++ src/store/modules/product/ProductState.ts
@@ -1,0 +1,5 @@
+export default interface ProductState {
+    cached: any;
+    products: {};
+  }
+
