+++ src/store/modules/product/actions.ts
@@ -1,0 +1,40 @@
+import { ActionTree } from 'vuex'
+import RootState from '@/store/RootState'
+import ProductState from './ProductState'
+import * as types from './mutation-types'
+import { ProductService } from '@/services/ProductService'
+import { hasError } from '@/utils'
+
+const actions: ActionTree<ProductState, RootState> = {
+
+  /**
+   * Fetch cached products
+   */
+  async fetchProducts ( { commit, state }, { productIds }) {
+    const cachedProductIds = Object.keys(state.cached);
+    const productIdFilter= productIds.reduce((filter: string, productId: any) => {
+      if (filter !== '') filter += ' OR '
+      // If product already exist in cached products skip
+      if (cachedProductIds.includes(productId)) {
+        return filter;
+      } else {
+        return filter += productId;
+      }
+    }, '');
+
+    // If there are no products skip the API call
+    if (productIdFilter === '') return;
+
+    const resp = await ProductService.fetchProducts({
+      "filters": ['productId: (' + productIdFilter + ')']
+    })
+    if (resp.status === 200 && !hasError(resp)) {
+      const products = resp.data.response.docs;
+      // Handled empty response in case of failed query
+      if (resp.data) commit(types.PRODUCT_ADD_TO_CACHED_MULTIPLE, { products });
+    }
+    // TODO Handle specific error
+    return resp;
+  },
+}
+export default actions;
