+++ src/store/modules/product/mutation-types.ts
@@ -1,0 +1,2 @@
+export const SN_PRODUCT = 'product'
+export const PRODUCT_ADD_TO_CACHED_MULTIPLE = SN_PRODUCT + '/ADD_TO_CACHED_MULTIPLE'
